refactor Discovery class
