+++ src/test/java/com/mascotapp/core/discoverer/DiscovererTest.java
@@ -1,7 +1,8 @@
-package com.mascotapp.core.discovery;
+package com.mascotapp.core.discoverer;
 
 import org.junit.jupiter.api.Test;
 
+import com.mascotapp.core.discoverer.Discoverer;
 import com.mascotapp.core.service.socialNetwork.SocialNetwork;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
@@ -10,13 +11,13 @@ import static org.junit.jupiter.api.Assertions.fail;
 import java.io.FileNotFoundException;
 import java.util.Set;
 
-public class MascotAppDiscoveryTest {
+public class DiscovererTest {
 
     @Test
     public void testDiscoverWithNonExistentPath() {
         String nonExistentPath = "nonexistent/directory";
         try {
-        	MascotAppDiscovery.discover(nonExistentPath);
+        	Discoverer.discover(nonExistentPath);
             fail("FileNotFoundException expected for a non-existent path");
         } catch (FileNotFoundException e) {
             // Exception expected
@@ -27,7 +28,7 @@ public class MascotAppDiscoveryTest {
     public void testDiscoverWithInvalidPath() {
         String invalidPath = "invalid path";
         try {
-        	MascotAppDiscovery.discover(invalidPath);
+        	Discoverer.discover(invalidPath);
             fail("IllegalArgumentException expected for an invalid path");
         } catch (IllegalArgumentException | FileNotFoundException e) {
             // Exception expected
@@ -37,7 +38,7 @@ public class MascotAppDiscoveryTest {
     @Test
     public void testDiscoverWithValidPathZeroImplementation() throws FileNotFoundException, IllegalArgumentException {
         String validPath = "src/test/resources/US2/ZeroImplementation";
-        Set<SocialNetwork> providers = MascotAppDiscovery.discover(validPath);
+        Set<SocialNetwork> providers = Discoverer.discover(validPath);
         assertEquals(providers.size(), 0);
     }
 }
